requirements.txt
Pillow>=10.0.0
PyGObject>=3.42.0
